+++ R/defaultPlotConfig.R
@@ -160,6 +160,8 @@
 #' @param return_interactive Whether to return an interactive plot.
 #' @param return_midparent Whether to return midparent values in the plot.
 #' @param override_many2many Whether to override many-to-many link logic.
+#' @param hints Optional hints to pass along to kinship2::autohint
+#' @param relation Optional relation to pass along to kinship2::pedigree
 #' @param debug Whether to enable debugging mode.
 #' @param ... Additional arguments for future extensibility.
 #' @return A named list of default plotting and layout parameters.
@@ -345,6 +347,9 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
                                  return_widget = FALSE,
                                  return_interactive = FALSE,
                                  return_midparent = FALSE,
+                                 # ---- Kinship2 Options ----
+                                 hints = NULL,
+                                 relation = NULL,
                                  # ---- Debugging Options ----
                                  debug = FALSE,
                                  override_many2many = FALSE,
@@ -446,9 +451,7 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
     # ---- Generation Scale Settings ----
     generation_height = generation_height,
     generation_width = generation_width,
-    ped_packed = ped_packed,
-    ped_align = ped_align,
-    ped_width = ped_width,
+
 
     # ---- Segment Drawing Options ----
 
@@ -558,6 +561,12 @@ getDefaultPlotConfig <- function(function_name = "getDefaultPlotConfig",
     return_widget = return_widget,
     return_interactive = return_interactive,
     return_midparent = return_midparent,
+    # ---- Kinship2 Options ----
+    ped_packed = ped_packed,
+    ped_align = ped_align,
+    ped_width = ped_width,
+    hints = hints,
+    relation = relation,
     # ---- Debugging Options ----
     override_many2many = override_many2many,
     debug = debug